% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_detect.R
\name{assoc_detect}
\alias{assoc_detect}
\title{Detecting associations between a discrete and a set of continuous features.}
\usage{
assoc_detect(
  data,
  marginals,
  target_inx,
  pred_inx,
  delta,
  mink_order = 1,
  alpha1 = 0.001,
  alpha2 = 0.1
)
}
\arguments{
\item{data}{Data set of observations, should be of class 'data.frame' with mixed-type data.}

\item{marginals}{Vector of row indices for marginal outliers detected. These are excluded from the analysis.}

\item{target_inx}{Index of target discrete variable, should be of class 'factor'.}

\item{pred_inx}{Indices of predictor continuous variables, should be of class 'numeric'.}

\item{delta}{Proportion of nearest neighbours considered. Should be a number in the range (0, 0.5].}

\item{mink_order}{Order of Minkowski distance. The default value of 1 returns the L1 norm (Manhattan distance).}

\item{alpha1}{Significance level of Kruskal Wallis H test used; default value is 1e-3.}

\item{alpha2}{Significance level for the chi-square goodness of fit test used for each class; default value is 1e-1.}
}
\value{
A vector of the predictor indices associated with the target discrete feature of interest. If NULL,
no association has been detected.
}
\description{
Detecting associations between a discrete and a set of continuous features.
}
\examples{
dt <- DOMID::gen_marg_joint_data(n_obs = 1000, n_disc = 5, n_cont = 10, n_lvls = 4,
                                           p_outs = 0.20, jp_outs = 0.80, assoc_target = 1,
                                           assoc_var = c(3, 4, 10), assoc_type = "quotient",
                                           seed_num = 1)
disc_scores <- DOMID::disc_scores(data = dt, disc_cols = c(1:5), alpha = 0.01, MAXLEN = 0)
cont_scores <- DOMID::cont_scores(data = dt, cont_cols = c(6:15), sample_size = 256,
                                  ntrees = 500, ndim = 0, max_depth = 100, seed_num = 1)
marginals <- DOMID::marg_outs_scores(data = dt, disc_cols = c(1:5), outscorediscdf = disc_scores[[2]],
                                     outscorecontdf = cont_scores, outscorediscdfcells = disc_scores[[3]],
                                     alpha = 0.01, rho = 0.2, epsilon = 0.02)
marginals <- unique(unlist(marginals))
DOMID::assoc_detect(data = dt, marginals = marginals, target_inx = 1, pred_inx = c(6:15), delta = 0.10)

}
