% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vecs_equal.R
\name{check_vecs_equal}
\alias{check_vecs_equal}
\title{Check if 2 vectors are the same}
\usage{
check_vecs_equal(vec1, vec2)
}
\arguments{
\item{vec1}{Vector 1}

\item{vec2}{Vector 2}
}
\value{
Sum of common elements of vec1 and vec2
}
\description{
Check if 2 vectors are the same
}
\examples{
check_vecs_equal(c(1,2), c(1,0))
}
