% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_scores.R
\name{cont_scores}
\alias{cont_scores}
\title{Continuous scores of outlyingness}
\usage{
cont_scores(
  data,
  cont_cols,
  sample_size = 256,
  ntrees = 500,
  ndim = 0,
  max_depth = 100,
  seed_num = 1
)
}
\arguments{
\item{data}{Data frame including continuous variables on which outlier detection will be implemented. Should be of class 'data.frame'.}

\item{cont_cols}{Column indices for discrete variables, these columns should be of class 'numeric'.}

\item{sample_size}{Sample size for Extended Isolation Forest algorithm, default choice is 256. If the data set includes less than sample_size observations, then the number of observations is used as sample_size instead.}

\item{ntrees}{Number of binary trees used in Extended Isolation Forest algorithm, default choice is 500.}

\item{ndim}{Number of dimensions on which the splits are made in Extended Isolation Forest algorithm, default choice of 0 corresponds to the number of continuous variables in the data. (Note: Setting ndim = 1 yields the original Isolation Forest algorithm.)}

\item{max_depth}{Maximum depth of binary trees used in Extended Isolation Forest algorithm, default choice is 100.}

\item{seed_num}{Seed number for reproducibility, default choice is 0.}
}
\value{
A data frame with 2 columns, the first corresponding to the observations and the second corresponding to their continuous scores of outlyingness.
}
\description{
Function used to calculate the continuous scores of outlyingness, based on the Extended Isolation Forest
algorithm of Hariri et al. (2019).
}
\examples{
cont_scores(data = iris, cont_cols = c(1:4))
cont_scores(data = iris, cont_cols = c(1:4), sample_size = 256, ntrees = 10, ndim = 1, max_depth = 10, seed_num = 100)
}
