% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fun.R
\name{dist_fun}
\alias{dist_fun}
\title{Weighted Minkowski distance matrix function}
\usage{
dist_fun(x, coeffs, degree)
}
\arguments{
\item{x}{Observations}

\item{coeffs}{Weights}

\item{degree}{Minkowski order}
}
\value{
Weighted Minkowski distance matrix
}
\description{
Weighted Minkowski distance matrix function
}
