% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_helper_fun.R
\name{distance_helper_fun}
\alias{distance_helper_fun}
\title{Weighted Minkowski distance between 2 observations}
\usage{
distance_helper_fun(x, y, w, deg)
}
\arguments{
\item{x}{Point (observation) 1}

\item{y}{Point (observation) 2}

\item{w}{Weights vector}

\item{deg}{Minkowski order}
}
\value{
Weighted Minkowski distance between x and y
}
\description{
Weighted Minkowski distance between 2 observations
}
