% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbow_angle.R
\name{elbow_angle}
\alias{elbow_angle}
\title{Calculate angle of elbow point in a decreasing convex function.}
\usage{
elbow_angle(vec, range)
}
\arguments{
\item{vec}{Vector of values - the values should be decreasing and the vector should include at least 4 values.}

\item{range}{Range vector - this corresponds to the x values corresponding to the vec values. Values should be unique and increasing, while the vector needs to be of length equal to vec.}
}
\value{
A numeric value of the elbow angle in degrees.
}
\description{
This function uses the kneedle algorithm for finding the knee of a decreasing convex function.
It then links the knee point to the lowest and highest points of the curve and calculates the
angle between these 2 line segements. If the 2 input arguments are very short vectors
(e.g. including just 5 elements) and no elbow is present, the angle value returned will be NA.
}
\examples{
elbow_angle(vec = sort(rnorm(10), decreasing = TRUE), range = c(1:10))
elbow_angle(vec = sort(rgamma(100, 1, 4), decreasing = TRUE), range = sort(rexp(100, 3)))
elbow_angle(vec = sort(rpois(20, 8.4), decreasing = TRUE), range = seq(0, 90, length.out = 20))

}
