% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_marg_joint_data.R
\name{gen_marg_joint_data}
\alias{gen_marg_joint_data}
\title{Generate mixed data with marginal & joint outliers}
\usage{
gen_marg_joint_data(
  n_obs,
  n_disc,
  n_cont,
  n_lvls,
  p_outs,
  jp_outs,
  assoc_target,
  assoc_vars,
  assoc_type = "linear",
  seed_num = 1
)
}
\arguments{
\item{n_obs}{Number of observations.}

\item{n_disc}{Number of discrete variables.}

\item{n_cont}{Number of continuous variables.}

\item{n_lvls}{Number of levels for discrete variables. If this is equal to an integer, all discrete variables will have the same number of levels. If some discrete variables have a different number of levels, a vector of length n_disc should be provided for the number of levels of each discrete level separately.}

\item{p_outs}{Proportion of observations in the data which are outliers (marginal or joint). This cannot exceed 0.2.}

\item{jp_outs}{Proportion of outliers which are joint. A value of 0 will return a data set with only marginal outliers, while a value of 1 will return a data set with only joint outliers.}

\item{assoc_target}{Indices of target discrete variables associated with sets of continuous features, on which joint outliers will be defined. This can be a single number if only one association exists in the data set, otherwise a vector indicating which discrete variables are associated with some continuous features.}

\item{assoc_vars}{Indices of sets of continuous variables associated with assoc_target. If there only exists one association in the data, this is a vector, otherwise it should be a list of vectors, the list having length equal to length(assoc_target).}

\item{assoc_type}{Type of association between discrete feature and a set of continuous variables; only "linear", "product" or "quotient" are supported. If assoc_vars is a vector of 1 element, argument is ignored. If multiple associations are present, a vector of association types should be provided for each association, or one type for all associations. The "linear" type cannot be given as input together with "product" or "quotient" due to a transformation that is applied on assoc_vars (except if the sets of assoc_vars are disjoint). Default is "linear".}

\item{seed_num}{Seed number (default value is 1).}
}
\value{
Data frame with mixed-type data and the specified number of observations and features of each type.
An additional column is added as an extra feature in the data set, indicating if an observation
is an inlier, a single marginal, a combined marginal or a joint outlier
(corresponding to values 0, 1, 2 and 3, respectively). Discrete variables are returned as factor class.
}
\description{
Function used to generate datasets with mixed-type variables and marginal & joint
outliers. The function returns a dataset of the dimensions provided with an additional column at the end,
stating whether each observation is an inlier, a single marginal, a combined marginal or a
joint outlier (corresponding to numbers 0, 1, 2 and 3, respectively).
Discrete variables are generated using quantile discretisation. The joint outliers are defined
by imposing an association between some of the discrete features and sets of continuous variables.
Only the linear, product and quotient designs are supported for now.
}
\examples{
data <- gen_marg_joint_data(n_obs = 1000, n_disc = 5, n_cont = 5, n_lvls = 3, p_outs = 0.05, jp_outs = 0.2, assoc_target = 1, assoc_vars = c(1, 2), assoc_type = "linear", seed_num = 1)
data <- gen_marg_joint_data(n_obs = 3000, n_disc = 4, n_cont = 8, n_lvls = c(3, 4, 3, 4), p_outs = 0.10, jp_outs = 0.6, assoc_target = 2, assoc_vars = c(1, 2, 7), assoc_type = "product", seed_num = 1)
data <- gen_marg_joint_data(n_obs = 5000, n_disc = 5, n_cont = 3, n_lvls = 4, p_outs = 0.20, jp_outs = 0.1, assoc_target = c(2, 5), assoc_vars = list(c(1, 2), c(2,3)), assoc_type = c("product", "quotient"), seed_num = 1)
}
