% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intv.R
\name{intv}
\alias{intv}
\title{Quantile discretisation of continuous vector}
\usage{
intv(vec, class)
}
\arguments{
\item{vec}{Vector to be discretised}

\item{class}{Number of levels}
}
\value{
Discretised vector (using quantile discretisation)
}
\description{
Quantile discretisation of continuous vector
}
\examples{
intv(vec = c(1:100), class = 5)
}
