% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_outs_lvls.R
\name{joint_outs_lvls}
\alias{joint_outs_lvls}
\title{Generate joint outliers based on a given association}
\usage{
joint_outs_lvls(data, level, value, quantiles_vec, out_label, target)
}
\arguments{
\item{data}{Data frame on which joint outliers will be introduced.}

\item{level}{Level of discrete features.}

\item{value}{Value of relationship.}

\item{quantiles_vec}{Vector of quantiles for given relationship.}

\item{out_label}{Index of label column, indicating which observation is outlying.}
}
\value{
New level, not agreeing to specified pattern,
}
\description{
Generate joint outliers based on a given association
}
