% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE_pred_1d.R
\name{KDE_pred_1d}
\alias{KDE_pred_1d}
\title{Predictions for 1-dimensional Kernel Density Estimation (w/ Gaussian kernel only)}
\usage{
KDE_pred_1d(t, xs, h)
}
\arguments{
\item{t}{New point to predict}

\item{xs}{Vector of points on which the density was evaluated}

\item{h}{Bandwidth for KDE}
}
\value{
Prediction for 1D Gaussian KDE
}
\description{
Predictions for 1-dimensional Kernel Density Estimation (w/ Gaussian kernel only)
}
