% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_lvl_fun.R
\name{new_lvl_fun}
\alias{new_lvl_fun}
\title{Generation of mild overlap between classes}
\usage{
new_lvl_fun(new_lvl, value, quantiles_vec)
}
\arguments{
\item{new_lvl}{New discrete level.}

\item{value}{Value of relationship.}

\item{quantiles_vec}{Vector of quantiles for given relationship.}
}
\value{
New discrete level for discrete observation.
}
\description{
Generation of mild overlap between classes
}
