% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_upper_fun.R
\name{u_upper_fun}
\alias{u_upper_fun}
\title{Function to determine the upper value of Theil's U, u_upper}
\usage{
u_upper_fun(x1, x2)
}
\arguments{
\item{x1}{Vector 1, consisting of factors}

\item{x2}{Vector 2, also consisting of factors}
}
\value{
Upper value of Theil's U, u_upper
}
\description{
Function to determine the upper value of Theil's U, u_upper
}
\examples{
u_upper_fun(as.factor(c(1,1,2)), as.factor(c(0,1,2)))
}
