% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg_outs.R
\name{marg_outs}
\alias{marg_outs}
\title{Detection of marginal outliers.}
\usage{
marg_outs(data, disc_cols, cont_cols)
}
\arguments{
\item{data}{Data frame of mixed-type data. Should be of class 'data.frame'.}

\item{disc_cols}{Column indices of discrete variables.}

\item{cont_cols}{Column indices of continuous variables.}
}
\value{
A list with 3 vectors; the vector "Discrete" includes the row indices
for the observations which are marginally outlying in the discrete space only (single
marginal outliers). The vector "Continuous" includes the row
indices for the observations which are marginally outlying in the continuous space only (only
single marginal outliers included). The vector "Combined" includes the row indices for
the combined marginal observations.
}
\description{
Function used to detect marginal outliers. The function only requires the column indices of the
discrete and the continuous variables and uses the functions disc_scores and cont_scores with their
default hyperparameter values to compute scores of outlyingness. The user can refer to the function
marg_outs_scores should they wish to detect the marginal outliers using discrete or continuous scores
that have been computed using different methods.
}
\examples{
dt <- gen_marg_joint_data(n_obs = 1000, n_disc = 5, n_cont = 5, n_lvls = 3, p_outs = 0.05, jp_outs = 0.2, assoc_target = c(1, 2), assoc_vars = list(c(1, 2), c(4,5)), assoc_type = c('linear', 'product'), seed_num = 1)
marg_outs(data = dt, disc_cols = c(1:5), cont_cols = c(6:10))

}
